refactor: Rename package title #1
