+++ msa-common/src/main/java/com/example/common/ApiResponse.java
@@ -1,6 +1,5 @@
-package com.example.commonlibrary;
+package com.example.common;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
