+++ msa-eureka/src/test/java/com/example/eureka/MsaEurekaApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.apigateway;
+package com.example.eureka;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class ApiGatewayApplicationTests {
+class MsaEurekaApplicationTests {
 
 	@Test
 	void contextLoads() {
