+++ msa-gateway/src/main/java/com/example/gateway/ApiGatewayApplication.java
@@ -1,4 +1,4 @@
-package com.example.apigateway;
+package com.example.gateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
